Refactor: Move paddle slider value indicator shape
